refactor : 토스 결제 승인요청을 feignClient로 변경

+++ src/main/java/com/yes25/yes255orderpaymentserver/persistance/domain/Payment.java
@@ -67,7 +67,7 @@ public class Payment {
         LocalDateTime requestedAt = LocalDateTime.parse((String) jsonObject.get("requestedAt"), formatter);
 
         return Payment.builder()
-            .paymentAmount(BigDecimal.valueOf((Long) jsonObject.get("totalAmount")))
+            .paymentAmount(BigDecimal.valueOf((Integer) jsonObject.get("totalAmount")))
             .paymentKey((String) jsonObject.get("paymentKey"))
             .approveAt(approvedAt)
             .requestedAt(requestedAt)
